backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.34
alembic==1.13.2
psycopg2-binary==2.9.9
pydantic==2.8.2
pydantic-settings==2.4.0
python-dotenv==1.0.1
authlib==1.3.1
requests==2.32.3
pandas==2.2.2
cryptography==43.0.1
python-multipart==0.0.9

# testing
pytest==8.3.2
pytest-asyncio==0.24.0
httpx==0.27.2

